requirements.txt
django==5.2.7
django-crispy-forms==1.14.0
pillow
